studentList added

+++ src/main/java/manas/project/attendance/entity/User.java
@@ -1,6 +1,5 @@
 package manas.project.attendance.entity;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -20,7 +19,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String firstName;
     private String lastName;
@@ -36,6 +35,5 @@ public class User {
                 .lastName(form.getLastName())
                 .build();
     }
-
 
 }
